change fields' type to uint32 and fix type error

+++ sprite8.go
@@ -12,20 +12,20 @@ import (
 // sprite8 is an 8-bit color sprite.
 type sprite8 struct {
 	r           io.ReaderAt
-	frameWidth  int
-	frameHeight int
-	frameCount  int
+	frameWidth  uint32
+	frameHeight uint32
+	frameCount  uint32
 	offsets     []uint32
-	width       int
-	height      int
+	width       uint32
+	height      uint32
 }
 
 func newSprite8(r io.ReaderAt, header spriteHeader) (*sprite8, error) {
 	sp := &sprite8{
 		r:           r,
-		frameWidth:  int(header.FrameWidth),
-		frameHeight: int(header.FrameHeight),
-		frameCount:  int(header.FrameCount),
+		frameWidth:  header.FrameWidth,
+		frameHeight: header.FrameHeight,
+		frameCount:  header.FrameCount,
 		offsets:     make([]uint32, header.FrameCount),
 	}
 	if err := binary.Read(&offsetedReader{r, 0x4c0}, binary.LittleEndian, &sp.offsets); err != nil {
@@ -45,33 +45,33 @@ func (sp *sprite8) ColorBits() int {
 }
 
 func (sp *sprite8) FrameWidth() int {
-	return sp.frameWidth
+	return int(sp.frameWidth)
 }
 
 func (sp *sprite8) FrameHeight() int {
-	return sp.frameHeight
+	return int(sp.frameHeight)
 }
 
 func (sp *sprite8) FrameCount() int {
-	return sp.frameCount
+	return int(sp.frameCount)
 }
 
 func (sp *sprite8) Width() int {
-	return sp.width
+	return int(sp.width)
 }
 
 func (sp *sprite8) Height() int {
-	return sp.height
+	return int(sp.height)
 }
 
 func (sp *sprite8) Frame(idx int) (image.Image, error) {
-	if idx < 0 || idx > sp.frameCount-1 {
+	if idx < 0 || idx > int(sp.frameCount-1) {
 		return nil, errors.New("frame index out of range")
 	}
-	img := image.NewPaletted(image.Rect(0, 0, sp.frameWidth, sp.frameHeight), sprite8Palette)
+	img := image.NewPaletted(image.Rect(0, 0, int(sp.frameWidth), int(sp.frameHeight)), sprite8Palette)
 	r := bufio.NewReader(&offsetedReader{sp.r, 0xbf4 + int64(sp.offsets[idx])})
-	for y := 0; y < sp.frameHeight; y++ {
-		for x := 0; x < sp.frameWidth; {
+	for y := uint32(0); y < sp.frameHeight; y++ {
+		for x := uint32(0); x < sp.frameWidth; {
 			b, err := r.ReadByte()
 			if err != nil {
 				return nil, errors.Wrap(err, "failed to read frame data")
@@ -82,11 +82,11 @@ func (sp *sprite8) Frame(idx int) (image.Image, error) {
 					return nil, errors.Wrap(err, "failed to read frame data")
 				}
 				for ; c > 0; c-- {
-					img.SetColorIndex(x, y, b)
+					img.SetColorIndex(int(x), int(y), b)
 					x++
 				}
 			} else {
-				img.SetColorIndex(x, y, b)
+				img.SetColorIndex(int(x), int(y), b)
 				x++
 			}
 		}
